feat: change app.tsx on router & create comp SearchPage
